Export everything from an internal module
…in particular the Doc type, for writing adaptors from the library to
others.
